Openearthtool/python/applications/wps/processes/requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
bzr==2.7.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.2.5
configobj==4.7.2
CouchDB==1.0
cycler==0.10.0
decorator==3.4.0
distro==1.4.0
futures==3.3.0
GeoAlchemy2==0.2.4
gunicorn==19.3.0
html5lib==0.999
httplib2==0.8
idna==2.8
jdcal==1.4.1
Jinja2==2.7.2
keyring==3.5
kiwisolver==1.1.0
launchpadlib==1.10.2
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
logutils==0.3.5
lxml==3.4.4
Mako==1.0.1
MarkupSafe==0.18
matplotlib==1.3.1
meld3==0.6.10
netCDF4==1.1.7.1
nose==1.3.7
numpy==1.16.4
oauth==1.0.1
oauthlib==3.0.2
-e svn+https://svn.oss.deltares.nl/repos/openearthtools/trunk/python/OpenEarthTools@11873#egg=OpenEarthTools
pandas==0.16.0
paramiko==1.10.7
Paste==1.7.5.1
PasteDeploy==1.5.2
pbr==5.4.1
Pillow==2.3.0
pkg-resources==0.0.0
psycopg2==2.8.3
pyasn1==0.1.9
pycairo==1.18.1
pycrypto==2.6.1
PyGObject==3.32.2
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2015.2
-e git+https://www.github.com/geopython/PyWPS.git@c9b6f644a7fbd89b383dbdc01cf683c5a6047a9f#egg=pywps
PyYAML==3.10
rainbow-logging-handler==2.2.2
requests==2.22.0
scipy==1.2.2
SecretStorage==2.0.0
Shapely==1.4.4
simplejson==3.16.0
simplekml==1.3.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.0.0
stevedore==1.30.1
subprocess32==3.5.4
supervisor==4.0.4
testresources==2.0.1
tornado==5.1.1
urllib3==1.25.3
virtualenv==16.7.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wadllib==1.3.3
waitress==1.3.0
zope.interface==4.6.0
